+++ .gitignore
@@ -9,4 +9,4 @@ profile.out
 *.vtk
 prof
 docs/_build
-
+cython/cbecke.c
